Fix lint/test issues
